Use size_t not unsigned int.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@34832 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ link-grammar/structures.h
@@ -305,8 +305,8 @@ struct X_table_connector_struct
 /* from string-set.c */
 struct String_set_s
 {
-    unsigned int size;       /* the current size of the table */
-    unsigned int count;      /* number of things currently in the table */
+    size_t size;       /* the current size of the table */
+    size_t count;      /* number of things currently in the table */
     char ** table;  /* the table itself */
 };
